requirements.txt
certifi==2019.11.28
chardet==3.0.4
curl==0.0.1
Django==2.1.15
djangorestframework==3.9.4
idna==2.8
importlib-metadata==1.4.0
path==13.1.0
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
urllib3==1.25.8
zipp==2.1.0
